services/pipecat-copilot/pyproject.toml
[project]
name = "pipecat-copilot"
version = "0.1.0"
description = "Agent Assist Copilot service using Pipecat framework"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[deepgram,elevenlabs]>=0.0.95",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "openai>=1.3.0",
    "google-generativeai>=0.3.0",
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

